+++ services/preview/eventgrid/mgmt/2021-06-01-preview/eventgrid/enums.go
@@ -579,6 +579,23 @@ func PossibleTopicTypeProvisioningStateValues() []TopicTypeProvisioningState {
 	return []TopicTypeProvisioningState{TopicTypeProvisioningStateCanceled, TopicTypeProvisioningStateCreating, TopicTypeProvisioningStateDeleting, TopicTypeProvisioningStateFailed, TopicTypeProvisioningStateSucceeded, TopicTypeProvisioningStateUpdating}
 }
 
+// TopicTypeSourceScope enumerates the values for topic type source scope.
+type TopicTypeSourceScope string
+
+const (
+	// TopicTypeSourceScopeAzureSubscription ...
+	TopicTypeSourceScopeAzureSubscription TopicTypeSourceScope = "AzureSubscription"
+	// TopicTypeSourceScopeResource ...
+	TopicTypeSourceScopeResource TopicTypeSourceScope = "Resource"
+	// TopicTypeSourceScopeResourceGroup ...
+	TopicTypeSourceScopeResourceGroup TopicTypeSourceScope = "ResourceGroup"
+)
+
+// PossibleTopicTypeSourceScopeValues returns an array of possible values for the TopicTypeSourceScope const type.
+func PossibleTopicTypeSourceScopeValues() []TopicTypeSourceScope {
+	return []TopicTypeSourceScope{TopicTypeSourceScopeAzureSubscription, TopicTypeSourceScopeResource, TopicTypeSourceScopeResourceGroup}
+}
+
 // Type enumerates the values for type.
 type Type string